requirements.txt
ai4bharat-transliteration==1.1.3
numpy
openpyxl
pandas
regex
requests
tqdm
huggingface_hub
